+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - [FIX] Clickable modifier ktx
 - [FIX] Refactor `share:receive` module
 - [REFACTOR] Fix detekt compose issues
+- [REFACTOR] Format markdown changelog
 
 # 1.4.0